Add `sample_size` to `no_std` bencher
In the `std` bencher, we accept a sample size, and hold on to all
intermediate benchmark values, whose buffers we allocate on demand.
On the other hand, we expect to use the `no_std` modules in a setting
where we can't allocate, and actually want to write our intermediate
results out to another box for analysis. So, we instead accept a pre-
allocated buffer to collect intermediate samples, and `Bencher::bench()`
accepts a ref to an output buffer.